fixed cast

+++ src/container/LinkedCellContainer.cpp
@@ -131,13 +131,13 @@ size_t LinkedCellContainer::index(Particle &p) {
     if (pos[0] >= domain[0]) {
         x_ind = mesh[0] - 1;
     } else {
-        x_ind = static_cast<size_t>(floor(pos[0] / rcutoff)) + 1;
+        x_ind = static_cast<int>(floor(pos[0] / rcutoff)) + 1;
     }
 
     if (pos[1] >= domain[1]) {
         y_ind = (mesh[1] - 1) * mesh[0];
     } else {
-        y_ind = static_cast<size_t>(floor(pos[1] / rcutoff)) * mesh[0] + mesh[0];
+        y_ind = static_cast<int>(floor(pos[1] / rcutoff)) * mesh[0] + mesh[0];
     }
     return x_ind + y_ind;
 }
